backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6
fastembed==0.4.2
anthropic==0.86.0
python-multipart==0.0.20
pydantic-settings==2.7.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
traceloop-sdk>=0.33.11
opentelemetry-exporter-otlp-proto-http>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
